+++ backend/routes/routes.js
@@ -11,7 +11,7 @@ router.post("/meal", (req, res) => {
   const n = new mealSchema(req.body);
   n.save();
   res.send("Received meal form");});
-  router.post("/register", (req, res) => {
+router.post("/register", (req, res) => {
     if (req.body.phno.length == 10) {
       Register.find({ email: req.body.email }, function (err, docs) {
         if (err) {
