+++ frontend/src/assets/food.png
