+++ frontend/src/components/meals/meal_style.scss
