+++ frontend/src/components/registerLogin/Register.scss
@@ -1,5 +1,5 @@
 body{
-    background-image: url("../assets/restaurant.png");
+    background-image: url("../../assets/restaurant.png");
 }
 .reg {
     align-self: center;
