+++ frontend/src/components/reservations/home.scss
